Improve locked profile honoring logic

+++ delete.php
@@ -54,7 +54,7 @@ if ($deleteall) {
     profile_helper::clear_min_duration_cache_for_reason($combinedreasons);
 
     // Delete all profile records.
-    $DB->delete_records(profile::TABLE, ['lockreason' => '']);
+    $DB->delete_records_select(profile::TABLE, $DB->sql_equal('lockreason', '?'), ['']);
     redirect($returnurl, get_string('allprofilesdeleted', 'tool_excimer'));
 }
 
@@ -63,13 +63,14 @@ if ($deleteid) {
     // Clears the profile metadata cache affected by this record deletion.
     $conditions = [
         'id' => $deleteid,
-        'lockreason' => '',
     ];
-    $profile = $DB->get_record(profile::TABLE, $conditions, 'request, reason');
-    $cache->delete($profile->request);
-    profile_helper::clear_min_duration_cache_for_reason($profile->reason);
-    // Deletes the profile record.
-    $DB->delete_records(profile::TABLE, $conditions);
+    $profile = $DB->get_record(profile::TABLE, $conditions, 'request, reason, lockreason');
+    if (empty($profile->lockreason)) {
+        $cache->delete($profile->request);
+        profile_helper::clear_min_duration_cache_for_reason($profile->reason);
+        // Deletes the profile record.
+        $DB->delete_records(profile::TABLE, $conditions);
+    }
     redirect($returnurl, get_string('profiledeleted', 'tool_excimer'));
 }
 
@@ -78,7 +79,6 @@ if ($filter) {
     $filtervalue = json_decode($filter, true);
     if (!is_null($filtervalue)) {
         // Clears the profile metadata caches affected by this filter.
-        $filtervalue['lockreason'] = '';
         $requests = $DB->get_records(profile::TABLE, $filtervalue, '', 'DISTINCT request');
         $reasons = $DB->get_records(profile::TABLE, $filtervalue, '', 'DISTINCT reason');
